+++ src/app/status/status.component.ts
@@ -33,16 +33,21 @@ export class StatusComponent implements OnInit {
   }
 
   getStatusSearch() {
+
+    if (this.statusSearch.nameContain == undefined || this.statusSearch.nameContain == ''){
+      this.notifier.notify('error', 'Um nome para a pesquisa deve ser preenchido.');
+      return;
+    }
+
     this.loading = true;
     this.checkService.getStatusSearch(window.localStorage["token"], this.statusSearch.nameContain).subscribe(res => {
       if(res.result == 'SUCCESS') {
-        this.loading = false;
         this.statusSearch.responseList = res.msgSaida;
         this.statusSearch.responseHasItem = true;
       } else {
-        this.loading = false;
         this.notifier.notify('error', res.error[0].message);
       }
+      this.loading = false;
     }, error => {
       if (error.status == 401) {
         this.router.navigate(['/login']);
@@ -51,19 +56,19 @@ export class StatusComponent implements OnInit {
       }
       this.loading = false;
     });
+
   }
 
   getStatusById() {
     this.loading = true;
     this.checkService.getStatusById(window.localStorage["token"], this.statusSave.id).subscribe(res => {
       if(res.result == 'SUCCESS') {
-        this.loading = false;
         this.statusSave.name = res.msgSaida[0].name;
         this.statusSave.id = res.msgSaida[0].id;
       } else {
-        this.loading = false;
         this.notifier.notify('error', res.error[0].message);
       }
+      this.loading = false;
     }, error => {
       if (error.status == 401) {
         this.router.navigate(['/login']);
@@ -75,16 +80,21 @@ export class StatusComponent implements OnInit {
   }
 
   saveStatus() {
+
+    if (this.statusSave == undefined || this.statusSave.name == ''){
+      this.notifier.notify('error', 'O nome do status deve ser informado.');
+      return;
+    }
+
     this.loading = true;
     this.checkService.saveStatus(window.localStorage["token"], this.statusSave).subscribe(res => {
       if(res.result == 'SUCCESS') {
-        this.loading = false;
         this.notifier.notify('success', res.msgSaida[0]);
         this.startPage();
       } else {
-        this.loading = false;
         this.notifier.notify('error', res.error[0].message);
       }
+      this.loading = false;
     }, error => {
       if (error.status == 401) {
         this.router.navigate(['/login']);
@@ -93,20 +103,25 @@ export class StatusComponent implements OnInit {
       }
       this.loading = false;
     });
+    
   }
 
   deleteStatus(statusIdToDelete) {
+
+    if (!confirm("Realmente deseja excluir o status de id " + statusIdToDelete + "?")) {
+      return;
+    }
+    
     this.loading = true;
     this.checkService.deleteStatus(window.localStorage["token"], statusIdToDelete).subscribe(res => {
       if(res.result == 'SUCCESS') {
-        this.loading = false;
         this.statusSave.name = res.msgSaida[0].name;
         this.statusSave.id = res.msgSaida[0].id;
         this.getStatusSearch();
       } else {
-        this.loading = false;
         this.notifier.notify('error', res.error[0].message);
       }
+      this.loading = false;
     }, error => {
       if (error.status == 401) {
         this.router.navigate(['/login']);
@@ -115,6 +130,7 @@ export class StatusComponent implements OnInit {
       }
       this.loading = false;
     });
+
   }
 
   edit = function (statusId) {
